Major restructuring to models
